Renamed listing route, controller
As book and listing information appear together, renamed files to clearly convey the purpose and functionality.
