+++ src/routes/bookListingRoute.js
@@ -1,7 +1,7 @@
 const express = require('express')
 
 // controller functions import
-const { getLenderListing, searchListings, addBookToListing,/*  updateListing ,*/ deleteListing, updateBookAndListing } = require('../controllers/listingController');
+const { getLenderListings, searchListings, addBookToListing,deleteListing, updateBookAndListing } = require('../controllers/bookListingController');
 
 // check authentication middleware import
 const checkAuth = require('../middlewares/checkAuth')
@@ -12,7 +12,7 @@ const router = express.Router()
 router.use(checkAuth)
 
 // get all listings of a lender
-router.get('/', getLenderListing )
+router.get('/', getLenderListings )
 
 // Search listings based on book title
 router.get("/search", searchListings)
@@ -20,11 +20,8 @@ router.get("/search", searchListings)
 // add listing to the database (if the book does not exist add that to the database first)
 router.post('/', addBookToListing )
 
-// update listing details
-/* router.patch('/:listingId', updateListing ) */
-
 // update book and listing details
-router.patch('/updateBookAndListing/:bookId/:listingId', updateBookAndListing);
+router.patch('/:bookId/:listingId', updateBookAndListing);
 
 // delete listing
 router.delete('/:listingId', deleteListing )
